Move CLI to separate namespace
