+++ src/Console/CLI.php
@@ -1,11 +1,13 @@
 <?php
 
-namespace KeepersTeam\Webtlo;
+namespace KeepersTeam\Webtlo\Console;
 
 use Exception;
-use splitbrain\phpcli\PSR3CLIv3;
+use KeepersTeam\Webtlo\ApplicationFactory;
+use KeepersTeam\Webtlo\Utils;
 use splitbrain\phpcli\Colors;
 use splitbrain\phpcli\Options;
+use splitbrain\phpcli\PSR3CLIv3;
 
 class CLI extends PSR3CLIv3
 {
