+++ webtlo.php
@@ -2,7 +2,7 @@
 
 require_once __DIR__ . DIRECTORY_SEPARATOR . 'vendor' . DIRECTORY_SEPARATOR . 'autoload.php';
 
-use KeepersTeam\Webtlo\CLI;
+use KeepersTeam\Webtlo\Console\CLI;
 
 $cli = new CLI();
 $cli->run();
